Add Call and call helpers

+++ cypress/support/resetdb.js
@@ -12,6 +12,7 @@ const Inspection = require("../../src/server/models/Inspection");
 const Notification = require("../../src/server/models/Notification");
 const Job = require("../../src/server/models/Job");
 const Client = require("../../src/server/models/Client");
+const Call = require("../../src/server/models/Call");
 var { seedUsers } = require("../fixtures/seedUsers");
 var seedSettings = require("../fixtures/seedSettings.json");
 var { seedNotes } = require("../fixtures/seedNotes");
@@ -22,6 +23,8 @@ const { seedSections } = require("../fixtures/seedSections");
 const { seedInspections } = require("../fixtures/seedInspections");
 const { seedJobs } = require("../fixtures/seedJobs");
 const { seedClients } = require("../fixtures/seedClients");
+const { seedCalls } = require("../fixtures/seedCalls");
+const calls = require("../../src/server/routes/validation/calls");
 
 // Grab the URI for the DB.
 const db = process.env.DEV_DB_URI;
@@ -158,10 +161,20 @@ Settings.deleteMany({}).then(() => {
                                         }  
                                         console.log("The jobs have been saved.");
 
-                                        Notification.deleteMany({}).then(() => {
-                                          console.log("Notifications have been cleared.");
-                                          console.log("Exiting DB Reset Script");
-                                          process.exit();
+                                        Call.deleteMany({}).then(() => {
+
+                                          Call.insertMany(seedCalls, (err, calls) => {
+                                            if(!jobs) {
+                                              console.error("There was an issue saving the jobs.");
+                                              return process.exit();
+                                            }  
+                                            console.log("The calls have been saved.");
+                                            Notification.deleteMany({}).then(() => {
+                                              console.log("Notifications have been cleared.");
+                                              console.log("Exiting DB Reset Script");
+                                              process.exit();
+                                            })
+                                          })
                                         })
                                       });
                                     });
